feat(flow): clean up tests
