+++ jina/enums.py
@@ -144,71 +144,6 @@ class LogVerbosity(BetterEnum):
     CRITICAL = 50
 
 
-class SocketType(BetterEnum):
-    """Enums for representing the socket type in a pod."""
-
-    PULL_BIND = 0
-    PULL_CONNECT = 1
-    PUSH_BIND = 2
-    PUSH_CONNECT = 3
-    SUB_BIND = 4
-    SUB_CONNECT = 5
-    PUB_BIND = 6
-    PUB_CONNECT = 7
-    PAIR_BIND = 8
-    PAIR_CONNECT = 9
-    ROUTER_BIND = 10
-    DEALER_CONNECT = 11
-    ROUTER_CONNECT = 13
-
-    @property
-    def is_bind(self) -> bool:
-        """
-        Check if this socket is using `bind` protocol.
-
-        :return: True if this socket is using `bind` protocol else False.
-        """
-        return self.value % 2 == 0
-
-    @property
-    def is_receive(self) -> bool:
-        """
-        Check if this socket is used for receiving data.
-
-        :return: True if this socket is used for receiving data else False.
-        """
-        return self.value in {0, 1, 4, 5}
-
-    @property
-    def is_pubsub(self):
-        """
-        Check if this socket is used for publish or subscribe data.
-
-        :return: True if this socket is used for publish or subscribe data else False.
-        """
-        return 4 <= self.value <= 7
-
-    @property
-    def paired(self) -> 'SocketType':
-        """
-        Get the paired SocketType.
-
-        :return: a paired SocketType.
-        """
-        return {
-            SocketType.PULL_BIND: SocketType.PUSH_CONNECT,
-            SocketType.PULL_CONNECT: SocketType.PUSH_BIND,
-            SocketType.SUB_BIND: SocketType.PUB_CONNECT,
-            SocketType.SUB_CONNECT: SocketType.PUB_BIND,
-            SocketType.PAIR_BIND: SocketType.PAIR_CONNECT,
-            SocketType.PUSH_CONNECT: SocketType.PULL_BIND,
-            SocketType.PUSH_BIND: SocketType.PULL_CONNECT,
-            SocketType.PUB_CONNECT: SocketType.SUB_BIND,
-            SocketType.PUB_BIND: SocketType.SUB_CONNECT,
-            SocketType.PAIR_CONNECT: SocketType.PAIR_BIND,
-        }[self]
-
-
 class FlowBuildLevel(BetterEnum):
     """
     The enum for representing a flow's build level.
