+++ scripts/get-all-test-paths.sh
@@ -24,7 +24,9 @@ else
     # will be executed in one runner
     declare -a comma_separated_tests=("tests/unit/peapods/runtimes/worker","tests/unit/peapods/runtimes/head","tests/unit/peapods/runtimes/request_handlers/","tests/unit/peapods/runtimes/gateway/grpc/test_grpc_gateway_runtime.py","tests/unit/peapods/runtimes/gateway/graph/test_topology_graph.py","tests/unit/peapods/test_networking.py","tests/unit/peapods/pods/test_pods.py","tests/unit/peapods/pods/test_scale.py","tests/unit/peapods/pods/test_pod_factory.py","tests/integration/runtimes/test_runtimes.py","tests/integration/peas/test_pea.py","tests/integration/peas/container/test_pea.py","tests/integration/pods/test_pod.py","tests/unit/peapods/peas/container/test_container_pea.py","tests/unit/peapods/peas/test_pea.py","tests/unit/peapods/stream/test_helper.py","tests/unit/peapods/stream/test_stream.py","tests/unit/types","tests/unit/peapods/pods/test_k8s_pod.py")
 
-    dest=( "${distributed_tests[@]}" "${comma_separated_tests[@]}" )
+    #flow tests
+    declare -a comma_separated_tests_flow=("tests/unit/flow-construct","tests/unit/flow-example","tests/unit/flow-orchestrate")
+    dest=( "${distributed_tests[@]}" "${comma_separated_tests[@]}" "${comma_separated_tests_flow[@]}" )
 
     printf '%s\n' "${dest[@]}" | jq -R . | jq -cs .
 fi
